chore: Move Protocol interface to a seprare package to avoid interface re-defninition
Up until now the `Protocol` interface was defined in both
`omaha/protocol.go` and `omaha/v3/protocol.go` to avoid dependency
cycle. That approach was suboptimal, as the exact same interface would
have to be maintained in two places which would be error-prone.
