+++ omaha/factory.go
@@ -4,19 +4,20 @@ import (
 	"fmt"
 
 	v3impl "github.com/brave/go-update/omaha/v3"
+	"github.com/brave/go-update/protocol"
 )
 
 // Factory creates protocol handlers for a specific version
 type Factory interface {
 	// CreateProtocol returns a Protocol implementation for the requested version
-	CreateProtocol(version string) (Protocol, error)
+	CreateProtocol(version string) (protocol.Protocol, error)
 }
 
 // DefaultFactory is the default implementation of Factory
 type DefaultFactory struct{}
 
 // CreateProtocol returns a Protocol implementation for the requested version
-func (f *DefaultFactory) CreateProtocol(version string) (Protocol, error) {
+func (f *DefaultFactory) CreateProtocol(version string) (protocol.Protocol, error) {
 	if !IsProtocolVersionSupported(version) {
 		return nil, fmt.Errorf("unsupported protocol version: %s", version)
 	}
